+++ nf_requirements/v1alpha1/dnn_types.go
